+++ lib/src/circuit_breaker/enums/fast_circuit_breaker_state.dart
@@ -1,0 +1,11 @@
+/// Represents the current state of the Circuit Breaker
+enum FastCircuitBreakerState {
+  /// Normal operation. Requests are allowed to pass through.
+  closed,
+
+  /// Circuit is open due to consecutive failures. Requests are blocked immediately.
+  open,
+
+  /// Testing mode. A limited number of requests are allowed to check if the underlying issue is resolved.
+  halfOpen,
+}